clean up igv drs code
